Revise project structure and make login page.
Added all of the changes for the project structure, meaning that it
should be good to go now. The instructions provided in README.md
make it so that anyone can set up the project, though they're definitely
not perfect. The login page only shows two fields and a button that
sends a message to the console log–which is only useful for checking to
see if angular is actually set up correctly.