Add database configuration and update README.md
This commit is the result of attempting to install and setup the project
on my own computer. A Flask configuration file was added that is
neccesary to run. MySQL-Python was abandoned for PyMySQL due to Python 3
compatibility. The 'submits' table schema was added to the database
initialization script. Finally additional instructions were included to
the README.