Add Bootstrap and minimal style to login page.
Login page now has Bootstrap and a small amount of style added to it.
Particularly the login box is now centered on the page and preceded
by a "Please log in" heading. Ideally in the future this would be
done under the Angular route templating system, but now we have a
not-so-ugly login screen.