Give create user form half-decent validation.
The create user form now validates the input. It does it as Angular as
possible, with ng-messages. Most of the validation occurs in the view,
though some in the SettingsController. ngMessages is a seprate module
that needed to be included to the project, but will be accessible from
any form that needs validation now.