First attempt at scoreboard timer
This commit adds the basic functionality of adding a timer to the
scoreboard. However, it has had embarassingly little testing. Also, the
code assumes that the client and server clocks match up, which may not
be the case. In the future, an offset should be added, or some other
solution to prevent time mismatch. Also, a recalibration every so often
would be a good idea as well.