Correct problem manager to match submission model
The submission model was altered to use a custom class instead of an
automapped one from SQLAlchemy. This change was not in the problem
manager and was causing it to break. This commit fixes that.

Also, a `DROP DATABASE IF EXISTS acm` was added to the top of the
acm.sql dump. this can be avoided in the future by creating a mysql dump
using the following:

$ mysqldump --databases acm --add-drop-database > acm.sql