(Almost) Completely refactor problems in database
The problems table was split into three tables: problems, problem_data,
and sample_cases. problems no longer contains the time limit of a
problem, and the original pid or problem was changed to lowercase and
made the shortname. Now pid's are auto-incrementing integer values.
problem_data holds the description, input/output description, and time
limit of a problem. sample_cases holds sample input/output test cases.
All are connected by the pid of a problem.

Now that problems are identified by integer values, the directories
containing test data (residing in setup/data.zip) have been renamed to
integers corresponding to the pid's in the database.

A known bug: The comp_problems table was also updated to reflect the new
pid's, but the test directories did not include the problem "The Problem
of Parity". It was removed from the database, but since it was in a
competition, it is still identified as "Parity" in the db. This will be
corrected in a future commit.