Start work for displaying a problem in HTML
Added new files and routing to start the skeleton for a template to
display problems in HTML. Currently utilizing the problem array from the
main controller and added some additional data to each element. For
future: should look to trimming that array down and maybe using
pagination on the problems page.