Refactor to be (mostly) pylint approved
problem_manager module now includes docstrings for the module and all
its functions. Also refactored some of the logic to be more
straightforward and reduce almost all of the pylint problems. A few
still remain (and one was supressed), particularly in the area of
importing modules. Not exactly sure how to fix (or if it needs to be
fixed), but it works.