Make problems more RESTful
This commit makes some slight changes to make the api a bit more RESTful
when dealing with problems. For instance, editing a problem uses the PUT
HTTP method. Also, the problem is always identified by the route, where
either shortname or numeric pid is acceptatble.

This commit also fixes test cases deleting whenever editing a problem,
and adds the foreign in the comp_problems table so that competition
scoreboards will display properly.