Add unit test for editing an existing problem
A slight modification had to be made to the commit_to_session() methods
for the problem models. Similar changes may need to be made for other
  problems in the future.

The problem was that the models.py module was importing session using a
`from ... import ...` to import the session variable, which in effect
created a new session and subsequently caused an exception in SQLAlchemy
(since the same object was attached to multiple sessions). To fix, the
session variable is now an optional parameter. The caller now supplies
the global session vairable, ensuring that there is only one. _There can
only be one._