Fix problems for PR
The docs were modified to fix a typo and include relative links for the
table of contents.

The initialize_database.sh script was corrected to read from the correct
SQL file, in addition to ensuring that the user had backed up the
database (or at least has the acm_test.sql file) and now ensures that
the test database gets created.

Lastly, the SQLAlchemy code for connecting to the test db was moved from
database.py to test.py. Now the test db will only be connected when
actual testing is being done.