WIP Runs but not all sockets work
A few dependencies were updated and the project now uses
Python 3 to initialize and run. Works so far as I can tell,
but not throughly tested. Sockets *almost* work. They can
send/receive, but updating the status of a problem never
gets received by the client. Unsure why, since messages are
received from other parts of the app just fine. WIll investigate,
my suspicion is that it has to do with our multithreading.