App now uses Python 3!
This commit finishes the upgrade to Python 3 and gets back
the original features of the old version. Most notably, threads
needed to be monkeypatched in order to be compatible
with eventlet/greenlet's threads and still remain asynchronous.

A few other slight changes needed to be made in order to still
function with Python 3. Notably:

- Integer division to calculate submission times on the
scoreboard was changed from `/` to `//`

- Calls to the API in unit tests had to be explicitly decoded
before loaded as a JSON object.

At it's current state, this commit passes all availible unit
tests for the project.