Change socket use, clean up Scoreboard html
Since Flasknado.emit() sent a message to *all* clients, it meant that
every single client would receive a new offset every time a client
requested the offset from the server. It seemed like overkill. Instead,
the websockets can now be used to request the time from the server.
Also, the time until the competition starts is now displayed in
Bootstrap's 'jumbotron' class and the table is hidden until the
competition starts. The only downside is that now teams are not visible
until the competition starts.