Extend Flasknado to register callbacks
Flasknado now allows for callbacks to be registered; these
are functions that will be executed when the server receives
a message of a particular type. Currently no support for
handling data passed with that message.

The client offset socket conversation was expaneded so that
only the client who sends the API request will receive the
server time again, instead of every client in the client pool.