Refactor the judge
There were a lot of changes made here, but here are the main ones:

* The Judge is now a class. This means that the properties of it are
simply set when it is created, and calling Judge.run() will start the
execution of the class.

* Similarly, calling Judge.run_threaded() will start the execution on a
new thread.

* The Flasknado code has been removed from the Judge model for
separations of concerns. It doesn't make sense to be doing network
operations from models.

* Due to run_threaded() already calling a new thread, the JudgeThread
class has been removed, as it only added unnecessary complexity to the
code. The logic for terminating a subprocess is entirely handled by the
while loop in judge_process().