Add ranking page and API route
This lays the groundwork for the ranklist page. The table
displays users ranked by number of problems solved. The
API excludes users who have not solved a single problem.
Still left to be done is the database refactoring that will make
ranking calculations _much_ more efficient, along with
hiding the raking table until the number crunching is
complete.