First attempt to Dockerize auacm app
The app now inludes a `Dockerfile` as well as `docker-compose.yml`.
These are used to create Docker images that can be run virtually to host
the app. With Docker and Docker Compose installed, run

`$ docker-compose up`

in the same directory as `docker-compose.yml`. This will download and
build the images, then run the app.

Caveats:
- The database runs in a MySQL container, built from the `acm.sql` in
  the `docker-data` directory. Not sure how (or if) persistence works
  between runs.
- Autoreload still occurs (the app is tied to the source code on the
  machine), but it takes about 30 seconds to complete. It's much faster
  to just restart the container manually with
    `$ docker-compose restart web`